B #6114: Fix NETTX and NETRX host monitoring (#2494)
(cherry picked from commit eecbfe351ca4f995a45b0e55fe61cb148a7f3aa6)

# Conflicts:
#	src/im_mad/remotes/dummy-probes.d/vm/monitor/monitor.rb

+++ src/host/HostMonitoringTemplate.cc
@@ -29,7 +29,7 @@ using namespace std;
 
 int CapacityMonitoring::from_template(const Template &tmpl)
 {
-    int value;
+    unsigned long value;
     if (tmpl.get("FREECPU", value))
     {
         add("FREE_CPU", value);
@@ -54,7 +54,7 @@ int CapacityMonitoring::from_template(const Template &tmpl)
 
 int SystemMonitoring::from_template(const Template &tmpl)
 {
-    int value;
+    unsigned long value;
     if (tmpl.get("CPUSPEED", value))
     {
         add("CPU_SPEED", value);
